Update UI
All filter windows are now positioned to the very right of the storage tab.
All filter windows will now close with the storage tab.
All filter windows now pause the game.
The close button for the edit filter dialog now closes the whole thing, and instead there is now a back button.
Adding a next-in-priority filter will now also open it in the same click.
The new filter dialog will no longer absorb all input.
Improved the feedback of a couple of buttons.
+ Multiple other small UI changes.